added user login auth
