+++ api/.env
@@ -1,1 +1,3 @@
 DB_FULLURI = mongodb+srv://favouradmin:KNrQpNWi5Vpf5ETj@favourcluster.5lkv2.mongodb.net/FavourDB?retryWrites=true&w=majority
+ACCESS_TOKEN_SECRET=a597ea48ce840a3d661bdf8913a3d9749ca43eccae96f30a451cfd74b95b00b6339fb94c5f577ac41204f69ec458d8fadf0adac4b272fb0b2736c1f69a8e86100f926a9fed4997aa39cb25cba54a9942
+REFRESH_TOKEN=a597ea48ce840a3d661bdf8913a3d974df0adac4b272fb0b2736c1f69a8e8610339fb94c5f577ac41204f69ec458d8fa0f926a9fed4997aa39cb25cba54a99429ca43eccae96f30a451cfd74b95b00b6
