+++ api/package.json
@@ -28,6 +28,7 @@
     "nodemon": "^2.0.4",
     "passport": "^0.4.1",
     "passport-jwt": "^4.0.0",
+    "redux": "^4.0.5",
     "validator": "^13.1.17"
   }
 }
