+++ client/src/components/pages/Profile.jsx
@@ -1,7 +1,17 @@
 import React from "react";
+import Leaderboard from "../shared/Leaderboard";
+import "./../../styles/Home.css";
 
-function Profile(props) {
-  return <div>Profile Page</div>;
+
+
+
+
+function Home(props) {
+  return (
+    <div class="center">
+      <p>Hello {this.state.user.name}!</p>
+    </div>
+  );
 }
 
-export default Profile;
+export default Home;
